requirements.txt
python-telegram-bot==21.5
requests==2.31.0
aiohttp==3.9.5
aiofiles==23.2.1
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.4.0
apscheduler==3.10.4
cachetools==5.3.3
fake-useragent==1.5.1
